% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_attributes.R
\name{add_attributes}
\alias{add_attributes}
\title{Add BugPhyzz attribute values as columns to a data frame}
\usage{
add_attributes(data, database)
}
\arguments{
\item{data}{A data frame containing abundance values.}

\item{database}{A data frame containing BugPhyzz attribute values.}
}
\value{
A data frame with new columns corresponding to BugPhyzz attribute values.
}
\description{
\code{add_attributes} adds columns to a data frame corresponding to attribute values from a BugPhyzz database.
}
\details{
The \strong{data frame} must have a column named "NCBI_ID" with NCBI taxids and a column named "Abundance" with relative abundance values.

The \strong{database} (downloaded from BugPhyzz) must have a column named "NCBI_ID" with NCBI taxids and a column named "Attribute_value."

NAs are converted to 0s.
}
\examples{
# Add cellular respiration attributes
df <- AsnicarF2017_genus[[1]]
db <- attribute(keyword = "aerophilicity")[[1]]

x <- add_attributes(df, db)
x

}
\seealso{
\code{\link{get_score}}

Other signature score functions: 
\code{\link{get_score}()}
}
\concept{signature score functions}
