% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_hierarchy.R
\name{add_hierarchy}
\alias{add_hierarchy}
\title{Add taxonomic rank names}
\usage{
add_hierarchy(df, y)
}
\arguments{
\item{df}{A data frame with a column named "NCBI_ID" containing taxids.}

\item{y}{A string indicating the required taxonomic rank. e.g. "phylum."}
}
\value{
Data frame with an additional column containing taxa names.
}
\description{
Add a column containing taxonomic rank names to a data frame.
}
\examples{
df <- AsnicarF2017_genus[[1]]
df <- add_hierarchy(df, "genus")
df
}
\seealso{
\code{\link{get_hierarchy}};
\code{\link[taxize]{classification}}

Other taxonomy functions: 
\code{\link{get_hierarchy}()}
}
\concept{taxonomy functions}
