% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AsnicarF2017_genus}
\alias{AsnicarF2017_genus}
\title{AsnicarF2017 genera cMD3}
\format{
A list of data frames (tibbles) with two columns each: "NCBI_ID" and "Abundance." Each data set corresponds to one sample.
\describe{
\item{MV_FEI1_t1Q14}{dim: 18  2}
\item{MV_FEI2_t1Q14}{dim: 12  2}
\item{MV_FEI3_t1Q14}{dim: 16 2 }
\item{MV_FEI4_t1Q14}{dim: 7 2}
\item{MV_FEI4_t2F15}{dim: 19  2}
\item{MV_FEI4_t2Q15}{dim: 24  2}
\item{MV_FEI5_t1Q14}{dim: 20  2}
\item{MV_FEI5_t2Q14}{dim: 35  2}
\item{MV_FEI5_t3Q15}{dim: 32  2}
\item{MV_FEM1_t1Q14}{dim: 33  2}
\item{MV_FEM2_t1Q14}{dim: 41  2}
\item{MV_FEM3_t1Q14}{dim: 45  2}
\item{MV_FEM4_t1Q14}{dim: 32  2}
\item{MV_FEM4_t2F15}{dim: 41  2}
\item{MV_FEM4_t2Q15}{dim: 57  2}
\item{MV_FEM5_t1Q14}{dim: 43  2}
\item{MV_FEM5_t2Q14}{dim: 27  2}
\item{MV_FEM5_t3Q15}{dim: 49  2}
\item{MV_MIM1_t1M14}{dim: 1 2}
\item{MV_MIM2_t1M14}{dim: 1 2}
\item{MV_MIM3_t1M14}{dim: 2 2}
\item{MV_MIM4_t1M14}{dim: 1 2}
\item{MV_MIM4_t2F15}{dim: 2 2}
\item{MV_MIM5_t1M14}{dim: 2 2}
\item{MV_MIM5_t2M14}{dim: 2 2}
\item{MV_MIM5_t3F15}{dim: 16  2}
}
}
\source{
\url{https://github.com/waldronlab/curatedMetagenomicData}
}
\usage{
AsnicarF2017_genus
}
\description{
A data set containing NCBI taxids of bacteria and relative abundance values at the genus level of AsnicarF2017 from the curatedMetagenomicData 3 package.
}
\keyword{datasets}
