% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute.R
\name{attribute}
\alias{attribute}
\title{Fetch physiological attribute data}
\usage{
attribute(keyword = "all")
}
\arguments{
\item{keyword}{a character vector of attribute names desired (see \code{\link{curationLinks}}). For the available
attributes, run bugphyzz::attribute_list()}
}
\value{
a large list of tidy data.frames
}
\description{
Fetch physiological attribute data
}
\examples{
x <- attribute()
head(x[[1]])

y <- attribute(c("gram", "oxygen"))
head(y[[1]])
}
