% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical_summary_stats.R
\name{attribute_catalog}
\alias{attribute_catalog}
\title{Display number of genera per phylum per attribute in physiology dataset}
\usage{
attribute_catalog(data, phys)
}
\arguments{
\item{data}{a tibble object, converted from the output by the \link{physiologies} function
and has been converted by the \link{fetch_ranks} function}

\item{phys}{a string vector of the physiology name}
}
\value{
a ggplot
}
\description{
Display number of genera per phylum per attribute in physiology dataset
}
\examples{
\dontrun{
physiologies <- physiologies()
aer <- physiologies[["aerophilicity"]] \%>\%
as_tibble()
aer_ranks <- fetch_ranks(aer)
attribute_catalog(aer_ranks, "aerophilicity")
}
}
\keyword{internal}
