% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_list.R
\name{attribute_list}
\alias{attribute_list}
\title{Fetch list of available physiological attribute names}
\usage{
attribute_list()
}
\value{
a character vector of physiological attribute names
}
\description{
Fetch list of available physiological attribute names
}
\examples{
x <- attribute_list()
}
