% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bergeys_abstracts.R
\name{bergeys_abstracts}
\alias{bergeys_abstracts}
\title{Fetch Bergeys abstracts
This function fetches Kelly Eckenrode's and Jonathan Ye's scraping of the Bergey's abstracts and returns a tidy data.frame}
\usage{
bergeys_abstracts()
}
\value{
A tidy data.frame
}
\description{
Fetch Bergeys abstracts
This function fetches Kelly Eckenrode's and Jonathan Ye's scraping of the Bergey's abstracts and returns a tidy data.frame
}
\examples{
x <- bergeys_abstracts()
dim(x)
}
