% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_summary_stats.R
\name{boxplot_per_phylum}
\alias{boxplot_per_phylum}
\title{Display boxplot per phylum}
\usage{
boxplot_per_phylum(data, phys, unit = "unit")
}
\arguments{
\item{data}{a tibble object, converted from the output by the \link{physiologies} function}

\item{phys}{a string vector of the physiology name}
}
\value{
a ggplot
}
\description{
Display boxplot per phylum
}
\examples{
\dontrun{
library(dplyr)
physiologies <- physiologies()
gt <- physiologies[["growth temperature"]] \%>\%
as_tibble()
boxplot_per_phylum(gt, "growth temperature")
}
}
\keyword{internal}
