% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugphyzz.R
\name{bpScores}
\alias{bpScores}
\title{Get bugphyzz scores from signature}
\usage{
bpScores(x, ...)
}
\arguments{
\item{x}{A bugphyzz signature}

\item{...}{Other arguments}
}
\value{
A named numeric vector.
}
\description{
Get bugphyzz scores from signature
}
