% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{checkColumnValues}
\alias{checkColumnValues}
\title{Check if values are valid across the columns of a bugphyzz dataset}
\usage{
checkColumnValues(x, dataset_name = NULL)
}
\arguments{
\item{x}{A dataset from bugphyzz, e.g. aerophilicity.}

\item{dataset_name}{Character vector of length 1; optional name for the dataset.}
}
\value{
Prints a message on screen if invalid values are found in one or more columns.
}
\description{
\code{checkColumnValues} compares the values in each column of a bugphyzz dataset with
the set of valid values or characters defined in the \code{valid_values} column in the template.tsv
file, and prints an error message to screen if invalid values are found in a column.
}
\examples{

\dontrun{

oxygen <- bugphyzz::phyisiologies(keyword = "aerophilicity")
bugphyzz::checkColumnValues(x = oxygen[[1]], dataset_name = names(oxygen)[1])#'

}

}
\seealso{
\code{\link{checkColumnValues}}; \code{\link{checkColumnValuesList}}
}
