% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{checkColumnValuesDF}
\alias{checkColumnValuesDF}
\title{Check column values in a data frame}
\usage{
checkColumnValuesDF(dataset, dataset_name = NULL)
}
\arguments{
\item{dataset}{A data frame.}

\item{dataset_name}{Character string of length 1; name of the dataset. Default is NULL.}
}
\value{
\code{\link{checkColumnValuesDF}} invisibly returns a list of error conditions (subclass
"invalid_column_values" or "invalid_column_class"), and it also prints an error message if
a column contains invalid values
}
\description{
\code{checkColumnValuesDF} applies the \code{\link{checkColumnValues}} function to the columns of
a single dataset.
}
\examples{

\dontrun{

# Example 1

dataset_with_bad_values <- data.frame(
  NCBI_ID = letters[1:10],
  Genome_ID = 1:10,
  Accession_number = 1:10,
  Taxon_name = letters[1:10],
  Attribute = 1:10,
  Attribute_value = letters[1:10],
  Attribute_source = 1:10,
  Evidence = 1:10,
  Confidence_interval = 1:10,
  Note = 1:10
)

err <- bugphyzz:::checkColumnValuesDF(dataset_with_bad_values)
err
err$NCBI_ID$bad_values

# Example 2

aerophilicity <- bugphyzz:::physiologies(keyword = "aerophilicity")[[1]]
aerophilicity_errors <- bugphyzz:::checkColumnValuesDF(aerophilicity)
aerophilicity_errors$Evidence$bad_values

}


}
\seealso{
\code{\link{checkColumnValues}};
\code{\link{checkColumnValuesDF}};
\code{\link{checkColumnValuesList}}

Other check functions: 
\code{\link{checkColumnValuesList}()},
\code{\link{checkColumnValues}()},
\code{\link{checkRequiredColumnsList}()},
\code{\link{checkRequiredColumns}()}
}
\concept{check functions}
