% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{checkColumnValuesList}
\alias{checkColumnValuesList}
\title{Check if values are valid across the columns of a list of bugphyzz datasets}
\usage{
checkColumnValuesList(x)
}
\arguments{
\item{x}{A named list of datasets in bugphyzz.}
}
\value{
Prints a message on screen if invalid values are found in one or more columns of each dataset.
}
\description{
\code{checkColumnValuesList} applies the \code{checkColumnValues} function to a list of dataframes
imported from the bugphyzz package. The \code{checkColumnValues} function prints a message if
columns don't have valid values as specified in the template.tsv file.
}
\examples{
\dontrun{

bp <- bugphyzz::physiologies(keyword = "all")
bugphyzz::checkColumnValuesList(bp)

}

}
\seealso{
\code{\link{checkColumnValues}}; \code{\link{checkColumnValuesList}}
}
