% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{checkRequiredColumns}
\alias{checkRequiredColumns}
\title{Check required columns in a single bugphyzz dataset}
\usage{
checkRequiredColumns(x, dataset_name = NULL)
}
\arguments{
\item{x}{A dataset from bugphyzz}

\item{dataset_name}{Character of length 1; name of dataset. Default is NULL.}
}
\value{
An error message if any of the required columns is missing or is not in the right order.
Otherwise, it returns a message indicating that required columns are present and in the right order.
}
\description{
\code{checkRequiredColumns} checks if the required columns (see \code{\link{requiredColumns}}) are
present and in the right order in a single bugphyzz dataset. If a required column is missing or
is not in the right order, it prints a message indicating which columns must be added or reordered.
}
\examples{

\dontrun{

oxygen <- bugphyzz::physiologies(keyword = "aerophilicity")
bugphyzz::checkRequiredColumns(x = oxygen[[1]], dataset_name = names(oxygen)[1])

}


}
\seealso{
\code{\link{requiredColumns}};
\code{\link{checkRequiredColumns}};
\code{\link{checkRequiredColumnsList}}
}
