% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{checkRequiredColumnsList}
\alias{checkRequiredColumnsList}
\title{Check required columns across a list of bugphyzz datasets}
\usage{
checkRequiredColumnsList(x)
}
\arguments{
\item{x}{A list of bugphyzz datasets.}
}
\value{
An error message if any of the required columns is missing or is not in the right order.
Otherwise, it returns a message indicating that the required columns are present and in the right order.
}
\description{
\code{checkRequiredColumnsList} applies the \code{\link{checkRequiredColumns}} function to a list of
bugphyzz datasets. If a required column is missing or is not in the right order, it prints a
message indicating which columns must be added or reordered.
}
\examples{

\dontrun{

bp = bugphyzz::physiologies(keyword = "all")
bugphyzz::checkRequiredColumnsList(bp)

}

}
\seealso{
\code{\link{requiredColumns}};
\code{\link{checkRequiredColumns}};
\code{\link{checkRequiredColumnsList}}
}
