% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physiologies.R
\name{curationLinks}
\alias{curationLinks}
\title{Show links to curation spreadsheets}
\usage{
curationLinks()
}
\value{
A data.frame with physiology names and URLs.
}
\description{
\code{curationLinks} returns a data.frame with the links the the spreadhseets
of the physiologies (both csv exports and source).
}
\examples{
bugphyzz:::curationLinks()
}
\keyword{internal}
