% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fattyAcidComposition.R
\name{customLinks}
\alias{customLinks}
\title{Custom Links}
\usage{
customLinks(keyword = "all")
}
\arguments{
\item{keyword}{a character vector of custom links desired.
Use "all" for all available custom links.}
}
\value{
A data frame with custom links.
}
\description{
\code{customLinks} returns a data frame with links to custom data sets not included
in the \code{physiologies} function.
}
\examples{
library(bugphyzz)
customLinks(keyword = "all")
}
