% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_summary_stats.R
\name{display_freq}
\alias{display_freq}
\title{Display frequency}
\usage{
display_freq(data, phys)
}
\arguments{
\item{data}{a tibble object, converted from the output by the \link{physiologies} function}

\item{phys}{a string vector of the physiology name}
}
\value{
a ggplot
}
\description{
Display frequency
}
\examples{
physiologies <- physiologies()
gt <- physiologies[["growth temperature"]] \%>\%
as_tibble()
display_freq(gt, "growth temperature")
}
\keyword{internal}
