% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{.appendLinks}
\alias{.appendLinks}
\title{Append links to error table}
\usage{
.appendLinks(x)
}
\arguments{
\item{x}{A data frame or tibble.}
}
\value{
A tibble.
}
\description{
\code{appendLinks} appends the links of each dataset with errors when
\code{table = TRUE} is set in the calls to
\code{\link{.checkRequiredColumnsList}} and
\code{\link{.checkColumnValuesList}}
}
\seealso{
\code{\link{.checkRequiredColumnsList}};
\code{\link{.checkColumnValuesList}}
}
\keyword{internal}
