% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{.attributes}
\alias{.attributes}
\title{Print valid attributes}
\usage{
.attributes()
}
\value{
A character vector with valid attribute names.
}
\description{
\code{\link{.attributes}} is a helper function for the
\code{\link{.checkColumnValues}} function. \code{\link{.attributes}} print
the valid attribute names as indicated in the attributes.tsv file.
}
\seealso{
\code{\link{.checkColumnValues}}
}
\keyword{internal}
