% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bergeysAbstracts.R
\name{.bergeysAbstracts}
\alias{.bergeysAbstracts}
\title{Fetch Bergeys abstracts}
\usage{
.bergeysAbstracts()
}
\value{
A tidy data.frame.
}
\description{
\code{.bergeyAbstracts} fetches Kelly Eckenrode's and Jonathan Ye's scraping
of the Bergey's abstracts and returns a tidy data.frame.
}
\examples{
bergeys <- bugphyzz:::.bergeysAbstracts()
head(bergeys)
}
\keyword{internal}
