% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{.checkColumnValues}
\alias{.checkColumnValues}
\title{Check column values}
\usage{
.checkColumnValues(dat, col, dat_name = NULL, quiet_success = TRUE)
}
\arguments{
\item{dat}{A data frame imported through the bugphyzz functions.}

\item{col}{A character string indicating the name of the column.}

\item{dat_name}{A character string indicating the name of the dataset.
Default is NULL.}

\item{quiet_success}{If FALSE, an error message is printed when no errors
are found. Default is TRUE.}
}
\value{
An error condition of subclass "invalid_column_values".
If \code{quiet_success} is FALSE and no errors were
found, it returns NULL and prints a message indicating that no errors were
found.
}
\description{
\code{.checkColumnValues} checks if a single column in a data frame contains
valid values according to the extdata/curation_template.tsv file.
}
\examples{

\dontrun{

x <- physiologies("aerophilicity")[[1]]
err <- tryCatch(
    error = function(e) e,
    bugphyzz:::.checkColumnValues(x, "Taxon_name")
)

}

}
\seealso{
\code{\link{.checkColumnValues}};
\code{\link{.checkColumnValuesDF}};
\code{\link{.checkColumnValuesList}}

Other check functions: 
\code{\link{.checkColumnValuesDF}()},
\code{\link{.checkColumnValuesList}()},
\code{\link{.checkRequiredColumns}()},
\code{\link{.checkRequiredColumnsDF}()},
\code{\link{.checkRequiredColumnsList}()}
}
\concept{check functions}
\keyword{internal}
