% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{.checkColumnValuesDF}
\alias{.checkColumnValuesDF}
\title{Check column values in a data frame}
\usage{
.checkColumnValuesDF(dat, dat_name = NULL)
}
\arguments{
\item{dat}{A data frame.}

\item{dat_name}{Character string indicating the name of the dataset.
Default is NULL.}
}
\value{
Invisibly returns a list of error
conditions (subclass "invalid_column_values"),
and it also prints an error message if a column contains invalid values.
If no errors are found, it returns NULL and a message indicating it.
}
\description{
\code{.checkColumnValuesDF} applies the \code{\link{.checkColumnValues}}
function to all of the columns of a single dataset. It's a tryCatch
wrapper of \code{\link{.checkColumnValues}}.
}
\examples{

\dontrun{

x <- physiologies("aerophilicity")[[1]]
err <- bugphyzz:::.checkColumnValuesDF(x)

}

}
\seealso{
\code{\link{.checkColumnValues}};
\code{\link{.checkColumnValuesDF}};
\code{\link{.checkColumnValuesList}}

Other check functions: 
\code{\link{.checkColumnValuesList}()},
\code{\link{.checkColumnValues}()},
\code{\link{.checkRequiredColumnsDF}()},
\code{\link{.checkRequiredColumnsList}()},
\code{\link{.checkRequiredColumns}()}
}
\concept{check functions}
\keyword{internal}
