% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{.checkColumnValuesList}
\alias{.checkColumnValuesList}
\title{Check column values in a list of bugphyzz datasets}
\usage{
.checkColumnValuesList(dats)
}
\arguments{
\item{list}{A list of bugphyzz datasets.}
}
\value{
Invisibly returns a list of error conditions (subclass
"invalid_column_values" or "invalid_column_class"), and it also prints
an error message if a column contains invalid values in any of the datasets
in the list.
}
\description{
\code{.checkColumnValuesList} applies the \code{.checkColumnValuesDF}
function to a list of bugphyzz datasets.
}
\examples{

\dontrun{

x <- physiologies()
list_of_errors <- bugphyz:::.checkColumnValuesList(x)
list_of_errors

}

}
\seealso{
\code{\link{.checkColumnValues}};
\code{\link{.checkColumnValuesDF}};
\code{\link{.checkColumnValuesList}}

Other check functions: 
\code{\link{.checkColumnValuesDF}()},
\code{\link{.checkColumnValues}()},
\code{\link{.checkRequiredColumnsDF}()},
\code{\link{.checkRequiredColumnsList}()},
\code{\link{.checkRequiredColumns}()}
}
\concept{check functions}
\keyword{internal}
