% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{.checkRequiredColumns}
\alias{.checkRequiredColumns}
\title{Check required columns}
\usage{
.checkRequiredColumns(dat, dat_name = NULL)
}
\arguments{
\item{dat}{A data frame from bugphyzz.}

\item{dat_name}{A character string indicating the name of the dataset.
Default is NULL.}
}
\value{
An error condition of subclass "required_columns_missing" or
"required_columns_misplaced". NULL and a message if no errors are found.
}
\description{
\code{.checkRequiredColumns} checks if the required columns
(see \code{\link{.requiredColumns}}) are
present and in the right order in a single bugphyzz dataset.
}
\examples{

\dontrun{

x <- physiologies("aerophilicity")[[1]]
err <- tryCatch(bugphyzz:::.checkRequiredColumns(x), error = function(e) e)
err

}

}
\seealso{
\code{\link{.requiredColumns}};
\code{\link{.stop_required_columns_missing}};
\code{\link{.stop_required_columns_misplaced}};
\code{\link{.checkRequiredColumnsDF}};
\code{\link{.checkRequiredColumnsList}}

Other check functions: 
\code{\link{.checkColumnValuesDF}()},
\code{\link{.checkColumnValuesList}()},
\code{\link{.checkColumnValues}()},
\code{\link{.checkRequiredColumnsDF}()},
\code{\link{.checkRequiredColumnsList}()}
}
\concept{check functions}
\keyword{internal}
