% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{.checkRequiredColumnsDF}
\alias{.checkRequiredColumnsDF}
\title{Check required columns in a single bugphyzz dataset}
\usage{
.checkRequiredColumnsDF(dat, dat_name = NULL)
}
\arguments{
\item{dat}{A data frame from bugphyzz.}

\item{dat_name}{A character string indicating the name of the dataset.
Default is NULL.}
}
\value{
Invisibly returns an error condition
("required_columns_missing" or "required_columns_misplaced" subclass).
It also returns an error message if any of the required columns is missing
or is not in the right order.
If no errors are found, a message indicating that the required
columns are present and in the right order is printed.
}
\description{
\code{.checkRequiredColumnsDF} checks if the required columns
(see \code{\link{.requiredColumns}}) are
present and in the right order in a single bugphyzz dataset.
If a required column is missing or is not in the right order,
it prints a message indicating which columns must be added or reordered.
}
\examples{

\dontrun{

x <- physiologies("aerophilicity")[[1]]
err <- tryCatch(bugphyzz:::.checkRequiredColumnsDF(x), error = function(e) e)
err

}

}
\seealso{
\code{\link{.requiredColumns}};
\code{\link{.checkRequiredColumnsDF}};
\code{\link{.checkRequiredColumnsDFList}};

Other check functions: 
\code{\link{.checkColumnValuesDF}()},
\code{\link{.checkColumnValuesList}()},
\code{\link{.checkColumnValues}()},
\code{\link{.checkRequiredColumnsDFList}()}
}
\concept{check functions}
\keyword{internal}
