% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{.checkRequiredColumnsList}
\alias{.checkRequiredColumnsList}
\title{Check required columns across a list of bugphyzz datasets}
\usage{
.checkRequiredColumnsList(list)
}
\arguments{
\item{list}{A list of bugphyzz datasets.}
}
\value{
Invisibly returns a list of error conditions
("required_columns_missing" or "required_columns_misplaced" subclasses),
and it also prints an error message for each error found.
If no errors are found, a message indicating that the required
columns are present and in the right order is printed.
}
\description{
\code{.checkRequiredColumnsList} applies the
\code{\link{.checkRequiredColumns}} function to a list of
bugphyzz datasets. If a required column (see \code{\link{.requiredColumns}})
is missing or is not in the right order in a dataset,
\code{.checkRequiredColumnsList} prints a message indicating which columns
are missing or must be reordered.
}
\examples{

\dontrun{

x <- physiologies()
list_of_errors <- bugphyzz:::.checkRequiredColumnsList(x)
list_of_erros$length

}

}
\seealso{
\code{\link{.requiredColumns}};
\code{\link{.checkRequiredColumns}};
\code{\link{.checkRequiredColumnsList}};

Other check functions: 
\code{\link{.checkColumnValuesDF}()},
\code{\link{.checkColumnValuesList}()},
\code{\link{.checkColumnValues}()},
\code{\link{.checkRequiredColumns}()}
}
\concept{check functions}
\keyword{internal}
