% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocFileCache.R
\name{.createTaxonomyTable}
\alias{.createTaxonomyTable}
\title{Create taxonomy table}
\usage{
.createTaxonomyTable()
}
\value{
A taxonomy table (*.RData object) saved in the cache, and a message
indicating that the taxonomy table has been saved to the cache.
}
\description{
\code{.createTaxonomyTable} extracts all of the unique NCBI IDs across
BugPhyzz and uses the \code{\link[taxize]{classification}} function to
retrieve the full taxonomy ranks from the NCBI database, and then saves it as
a table in the user's cache with the \code{BiocFileCache} package. The
cache can be accessed with the \code{\link{.getCache}} function.
}
\details{
Two versions of the taxonomy table are saved in a list object (.RData),
one version with taxonomy names and another with NCBI IDs. See the
\code{\link{taxonomyTable}} function to know how to access each of them.
}
\examples{

\dontrun{
library(bugphyzz)
.createTaxonomyTable()

}

}
