% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{.err_list_to_table}
\alias{.err_list_to_table}
\title{Error list to table}
\usage{
.err_list_to_table(err)
}
\arguments{
\item{err}{List of errors.}
}
\value{
A tibble.
}
\description{
\code{.err_list_to_table} converts a list of errors generated with the
\code{\link{.checkColumnValuesList}} function into a tibble.
}
\seealso{
\code{\link{.checkColumnValuesList}}
}
\keyword{internal}
