% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{.errListToTable}
\alias{.errListToTable}
\title{Error list to table}
\usage{
.errListToTable(err)
}
\arguments{
\item{err}{List of errors.}
}
\value{
A tibble.
}
\description{
\code{.errListToTable} converts a list of errors generated with the
\code{\link{.checkColumnValuesList}} function into a tibble.
}
\seealso{
\code{\link{.checkColumnValuesList}}
}
\keyword{internal}
