% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{.errorHandler}
\alias{.errorHandler}
\title{Error handler}
\usage{
.errorHandler(FUN, ...)
}
\arguments{
\item{FUN}{A check function.}

\item{...}{Arguments passed to the check function indicated with the FUN
argument.}
}
\value{
An object of class "error", "conditiion" and variable subclass, e.g.,
"invalid_column_values!.
}
\description{
\code{.errorHandler} implements tryCatch to capture curation errors and their
metadata.
}
\keyword{internal}
