% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bergeys_textmine_functions.R
\name{.extractDuplicate}
\alias{.extractDuplicate}
\title{Function retrieves bacteria that need manual curation}
\usage{
.extractDuplicate(df)
}
\description{
Also writes to csv
}
\keyword{internal}
