% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.getCache}
\alias{.getCache}
\title{BugPhyzz's cache}
\usage{
.getCache()
}
\description{
\code{getCache} starts or loads the local cache. The location of the path is
determined by \code{tools::R_user_dir(package = "curatedMetagenomicDataClient", which = "cache")}.
}
\keyword{internal}
