% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physiologies2.R
\name{.importPhysiology}
\alias{.importPhysiology}
\title{Import physiology}
\usage{
.importPhysiology(x)
}
\arguments{
\item{x}{A row from the output of \code{curationLinks()}.}
}
\value{
A data frame.
}
\description{
\code{.importPhysiology} imports a physiology directly from one row of the
\code{curationLinks()} output.
}
\keyword{internal}
