% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physiologies2.R
\name{.importRange}
\alias{.importRange}
\title{Import range}
\usage{
.importRange(link)
}
\arguments{
\item{link}{A character string. Url to the spreadsheet.}
}
\value{
A data.frame.
}
\description{
\code{.importRange} imports a dataset labeled with "range" in
\code{curationLinks}.
}
\keyword{internal}
