% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physiologies.R
\name{.modifyRange}
\alias{.modifyRange}
\title{Modify attributes of type range}
\usage{
.modifyRange(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A data.frame.
}
\description{
\code{.modifyRange} imports a dataset labeled with "range" in
}
\keyword{internal}
