% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{.requiredColumns}
\alias{.requiredColumns}
\title{Required columns}
\usage{
.requiredColumns(attr_type)
}
\arguments{
\item{attr_type}{Attribute type. A character string. Options: logical,
numeric, character, range.}
}
\value{
A character vector with mandatory column names (required columns).
}
\description{
\code{.requiredColumns} prints the names and order of the columns that must
be present in all of the bugphyzz data sets.
}
\examples{

\dontrun{
bugphyzz:::.requiredColumns()
}

}
\seealso{
\code{\link{.checkRequiredColumns}}
}
\keyword{internal}
