% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bergeys_textmine_functions.R
\name{.standardize}
\alias{.standardize}
\title{Function replaces synonyms with standard title}
\usage{
.standardize(df, standard, synonyms = standard)
}
\description{
Function replaces synonyms with standard title
}
\examples{

\dontrun{

temp <- bugphyzz:::.standardize(
    dat, "gram", words[[1]], colnames(words)[[1]]
)

}

}
\keyword{internal}
