% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_functions.R
\name{.stop_custom}
\alias{.stop_custom}
\title{Custom function for helping generate stop conditions in the bugphyzz package}
\usage{
.stop_custom(subclass, message, call = NULL, ...)
}
\arguments{
\item{subclass}{Character vector of length 1; name of custom error condition.}

\item{message}{Character vector of length 1; an informative message.}

\item{call}{Include or not the function call that generates the error; e.g. \code{sys.call(-1)}.}

\item{...}{Any other argument useful to identify the source of the error and how to fix it.}
}
\value{
Error condition. Object of class "error", "condition", and a custom error class name (subclass).
}
\description{
\code{.stop_custom} is a helper function for generating custom stop conditions. This piece of
code was taken from Hadley Wickham's Advanced R book, 2nd edition
(\url{https://adv-r.hadley.nz/conditions.html#custom-conditions}).
}
