% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_functions.R
\name{.stop_invalid_column_class}
\alias{.stop_invalid_column_class}
\title{Stop condition for invalid column class (in Attribute_value column)}
\usage{
.stop_invalid_column_class(col_class, dataset_name = NULL, ...)
}
\arguments{
\item{col_class}{Character vector of length 1; name of invalid class.}

\item{dataset_name}{Character vector of length 1; name of the dataset.}

\item{...}{Any other argument useful to identify the source of the error and how to fix it.}
}
\value{
Error condition. Object of class: "invalid_column_class", "error", "condition".
}
\description{
\code{.stop_invalid_column_class} generates an error condition if a column (Attribute_value)
is not logical or numeric.
}
