% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{.stop_invalid_column_values}
\alias{.stop_invalid_column_values}
\title{Stop condition for invalid values in a column of a bugphyzz dataset}
\usage{
.stop_invalid_column_values(col, n_rows, dat_name = NULL, values = NULL, ...)
}
\arguments{
\item{col}{Character string containing a single column name.}

\item{n_rows}{Integer of length 1; number of rows with invalid values.}

\item{dat_name}{Character string indicating the dataset name.}

\item{values}{Optional.
Invalid values in column. Only the first three will be printed.}

\item{...}{Any other argument useful to identify the source of the error
and/or how to fix it.}
}
\value{
Error condition.
Object of class "invalid_column_names", "error", "condition".
}
\description{
\code{.stop_invalid_column_values} generates an error condition if a column
contains invalid values. This function should be used within the
\code{\link{.checkColumnValues}} function.
}
\seealso{
\code{\link{.stop_custom}};
\code{\link{.stop_invalid_column_values}};
\code{\link{.stop_invalid_column_class}};
\code{\link{.stop_uncataloged_column}};
\code{\link{.checkColumnValues}}

Other custom stop functions: 
\code{\link{.stop_custom}()},
\code{\link{.stop_invalid_column_class}()},
\code{\link{.stop_required_columns_misplaced}()},
\code{\link{.stop_required_columns_missing}()},
\code{\link{.stop_uncataloged_column}()}
}
\concept{custom stop functions}
\keyword{internal}
