% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_functions.R
\name{.stop_required_columns_misplaced}
\alias{.stop_required_columns_misplaced}
\title{Stop condition for misplaced required columns in a bugphyzz dataset}
\usage{
.stop_required_columns_misplaced(cols, dataset_name = NULL, ...)
}
\arguments{
\item{cols}{Character vector; names of missing columns.}

\item{dataset_name}{Character vector of length 1; dataset name.}

\item{...}{Any other argument useful to identify the source of the error and how to fix it.}
}
\value{
Error condition. Object of class: "required_columns_misplaced", "error", "condition".
}
\description{
\code{.stop_required_columns_misplaced} generates an error condition if one or more of the
required columns is not in the right order in a bugphyzz dataset. The order of required columns
can be printed with the \code{validHeaders} function.
}
