% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_functions.R
\name{.stop_required_columns_missing}
\alias{.stop_required_columns_missing}
\title{Stop condition for missing required columns in a bugphyzz dataset}
\usage{
.stop_required_columns_missing(cols, dataset_name, ...)
}
\arguments{
\item{cols}{Character vector; names of missing columns.}

\item{dataset_name}{Character vector of length 1; dataset name.}

\item{...}{Any other argument useful to identify the source of the error and how to fix it.}
}
\value{
Error condition. Object of class "required_columns_missing", "error", "condition".
}
\description{
\code{.stop_required_columns_missing} generates an error condition if one or more of the
required columns is missing from a bugphyzz dataset. The required columns can be printed with the
\code{validHeaders} function.
}
