% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{.stop_required_missing_row_value}
\alias{.stop_required_missing_row_value}
\title{Stop condition for missing and required row value in a bugphyzz dataset}
\usage{
.stop_required_missing_row_value(column, rows, dat_name = NULL, ...)
}
\arguments{
\item{column}{A character vector of the column name where data is missing.}

\item{rows}{An integer vector containing the position of missing row data.}

\item{dat_name}{A character string indicating the name of the dataset.}

\item{...}{Any other argument useful to identify the source of the error
and/or how to fix it.}
}
\value{
Error condition.
Object of class "required_missing_row_value", "error", "condition".
}
\description{
\code{.stop_required_missing_row_value} generates an error condition of
class "required_missing_row_value". This function should be used within the
\code{\link{.checkRequiredColumns}} function.
}
\seealso{
\code{\link{.stop_custom}};
\code{\link{.stop_required_columns_missing}};
\code{\link{.stop_required_columns_misplaced}};
\code{\link{.checkRequiredColumns}}

Other custom stop functions: 
\code{\link{.stop_custom}()},
\code{\link{.stop_invalid_column_values}()},
\code{\link{.stop_required_columns_misplaced}()},
\code{\link{.stop_required_columns_missing}()},
\code{\link{.stop_uncatalogued_column}()}
}
\concept{custom stop functions}
\keyword{internal}
