% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_functions.R
\name{.stop_uncatalogued_column}
\alias{.stop_uncatalogued_column}
\title{Stop condition for a column not catalogued in the template file}
\usage{
.stop_uncatalogued_column(col, dataset_name = NULL, ...)
}
\arguments{
\item{col}{Character vector of length 1; column name.}

\item{dataset_name}{Character vector of length 1; dataset name.}

\item{...}{Any other argument useful to identify the source of the error and how to fix it.}
}
\value{
Error condition. Object of class: "invalid_column_class", "error", "condition".
}
\description{
Stop condition for a column not catalogued in the template file
}
