% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bergeys_textmine_functions.R
\name{.sub_word}
\alias{.sub_word}
\title{Function modifies the bergey's df and substitutes words}
\usage{
.sub_word(df, standard, synonyms = standard)
}
\description{
Function modifies the bergey's df and substitutes words
}
\keyword{internal}
