% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{.template}
\alias{.template}
\title{Generates a template for a bugphyzz dataset}
\usage{
.template(x)
}
\arguments{
\item{x}{A data.frame or tibble; a dataset from bugphyzz.}
}
\value{
A tibble with column names and valid values for a given bugphyzz dataset.
}
\description{
Generates a template for a bugphyzz dataset
}
