% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_valid_headers.R
\name{expect_valid_headers}
\alias{expect_valid_headers}
\title{Expect that headers are valid in a single data frame}
\usage{
expect_valid_headers(x, y = "current")
}
\arguments{
\item{x}{A data frame from the bugphyzz database.}

\item{y}{A character string indicating the name of the data set in bugphyzz.
This name is custom, not taken from the data set itself.}
}
\value{
Error message if the headers are not valid or are not in the correct
order. Otherwise, it returns nothing.
}
\description{
\code{expect_valid_headers} tests if the first mandatory columns of a
bugphyzz data set have valid header names and are in the correct order.
}
\examples{
# Example 1
\dontrun{
library(bugphyzz)
datasets <- curationLinks()[, c("physiology", "link")]
link1 <- datasets[1, "link"]
df1 <- read.csv(link1)
expect_valid_headers(df1) # Error expected

# Example 2
library(bugphyzz)
link2 <- datasets[datasets[["physiology"]] == "aerophilicity", "link"]
df2 <- read.csv(link2)
expect_valid_headers(df2) # No error expected
}
}
