% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_valid_headers.R
\name{expect_valid_headers_all}
\alias{expect_valid_headers_all}
\title{Expect that headers are valid in a list of data frames}
\usage{
expect_valid_headers_all(x)
}
\arguments{
\item{x}{A named list of valid bugphyzz data frames.}
}
\value{
An error message per each data set that does not pass the test.
Otherwise, no output is returned.
}
\description{
Expect that headers are valid in a list of data frames
}
\examples{
\dontrun{
library(bugphyzz)
datasets <- curationLinks()[, c("physiology", "link")]
database <- vector("list", nrow(datasets))
for (i in seq_along(database)) {
  names(database)[i] <- datasets[i, "physiology"]
  database[[i]] <- read.csv(datasets[i, "link"])
}
expect_valid_headers_all(database)
}
}
