% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fattyAcidComposition.R
\name{fattyAcidComposition}
\alias{fattyAcidComposition}
\title{function for shaping fatty acid composition from Google Sheets using tidyR}
\usage{
fattyAcidComposition()
}
\arguments{
\item{keyword}{a character vector of custom links desired. For the available
custom links, run bugphyzz::custom_link(). Use "all" for all available custom links.}
}
\value{
a tidy data frame of fattyacidComposition from customlinks.tsv
}
\description{
function for shaping fatty acid composition from Google Sheets using tidyR
}
\examples{
fattyAcidComposition()
x <- fattyAcidComposition()
View(x)

}
