% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSignatures.R
\name{fetchBugs}
\alias{fetchBugs}
\title{Fetch bugs based on taxid and corresponding attribute}
\usage{
fetchBugs(attribute_name, dat, taxids = "Taxon_name")
}
\arguments{
\item{attribute_name}{name of attribute (aerobic, anaerobic, etc.))}

\item{dat}{a physiology data.frame}

\item{taxids}{identification of the bugs that will have the attribute value}
}
\value{
a character vector of taxids
}
\description{
Fetch bugs based on taxid and corresponding attribute
}
\examples{
physiology_list <- bugphyzz::physiologies(keyword = "aerophilicity")
df <- physiology_list[[1]]
x <- fetchBugs(attribute_name = "aerobic", dat = df, taxids = "Taxon_name")
head(x)
}
