% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_NCBI.R
\name{fill_NCBI}
\alias{fill_NCBI}
\title{Fill in NCBI IDs where the cells have NA}
\usage{
fill_NCBI(data)
}
\arguments{
\item{data}{A dataframe from the \link{physiologies} function.}
}
\value{
A dataframe.
}
\description{
\code{fill_NCBI} fills NA cells in the NCBI_ID column with taxids according
to the Taxon_name column.
}
\examples{
\dontrun{
aer <- physiologies("aerophilicity")[[1]]
aer_NCBI <- fill_NCBI(aer)
}
}
\keyword{internal}
