% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_attributes_abundance.R
\name{get_attributes_abundance}
\alias{get_attributes_abundance}
\title{Get summarized abundance values per attribute}
\usage{
get_attributes_abundance(data, database)
}
\arguments{
\item{data}{A data frame where the first column is named "NCBI_ID" and the
remaining columns contain abundance values per sample.}

\item{database}{A data set from the \code{bugphyzz} package, e.g. aerophilicity.}
}
\value{
A data frame (tibble) containing the summarized abundance values per
bugphyzz attribute per sample.
}
\description{
\code{get_abundance_attributes} takes a data frame in which the first column
contains NCBI taxids and the remaining columns contain abundance values
(one column per sample).
}
\examples{
library(bugphyzz)

aerophilicity <- physiologies(keyword = "aerophilicity")[[1]]

fpath <- system.file(path = "extdata", file = "AsnicarF_2017_genus.tsv", package = "bugphyzz")
asnicarf2017 <- readr::read_tsv(fpath)

attributes_abundance <- get_attributes_abundance(asnicarf2017, aerophilicity)
attributes_abundance

}
