% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_hierarchy.R
\name{get_hierarchy}
\alias{get_hierarchy}
\title{Get taxinomic rank names from an NCBI taxid}
\usage{
get_hierarchy(x, y)
}
\arguments{
\item{x}{A double, integer, or character indicating a valid NCBI taxid.}

\item{y}{A string indicating the taxonomic rank, e.g. "phylum."}
}
\value{
A string with the name of the taxa.
}
\description{
This function is a wrapper around \code{taxize::classification}. It takes a NCBI taxid and a taxonomic rank and returns the name of the taxa at the specified taxonomic rank. Valid inputs for taxonomic rank (for Bacteria): "superkingdom", "phylum", "class", "order", "family", "genus", "species."
}
\examples{

get_hierarchy(906, "phylum")
get_hierarchy(906, "genus")
}
\seealso{
\code{\link{add_hierarchy}};
\code{\link[taxize]{classification}}

Other taxonomy functions: 
\code{\link{add_hierarchy}()}
}
\concept{taxonomy functions}
