% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_score.R
\name{get_score}
\alias{get_score}
\title{Get signature scores from an abundance/attributes matrix}
\usage{
get_score(x)
}
\arguments{
\item{x}{A data frame. Output from the  \code{add_attributes} function.}
}
\value{
A data frame of two columns: "Attributes" and "Score."
}
\description{
\code{get_score} calculates a signature score for each attribute value (one attribute per column) by adding up the relative abundance values. Since the abundance values are added up by column, the values are always between 0 and 100.
}
\examples{
df <- AsnicarF2017_genus[[1]]
db <- attribute(keyword = "oxygen")[[1]]

x <- add_attributes(df, db)
x

y <- get_score(x)
y
}
\seealso{
\code{\link{add_attributes}}

Other signature score functions: 
\code{\link{add_attributes}()}
}
\concept{signature score functions}
