% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy_functions.R
\name{get_taxonomy_ranks}
\alias{get_taxonomy_ranks}
\title{Get Taxonomy Ranks from NCBI TaxIDs}
\usage{
get_taxonomy_ranks(x)
}
\arguments{
\item{x}{A numeric vector of valid NCBI taxids.}
}
\value{
A tibble in which each row contains the input NCBI IDs and each column contains
the taxonomy names of the following ranks: kingdom, phylum, class, order,
family, genus, species.
}
\description{
\code{get_taxonomy_ranks} retrieves taxonomy names for input NCBI taxids at
different taxonomy ranks.
}
\examples{
library(bugphyzz)
taxids <- c(1578, 745, 562, 2)
rank_names <- get_taxonomy_ranks(taxids)


}
