% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugphyzz.R
\name{getBugAnnotations}
\alias{getBugAnnotations}
\title{Get Bug Annotations}
\usage{
getBugAnnotations(x, bp = importBugphyzz(), tax.id.type)
}
\arguments{
\item{x}{A valid NCBI ID or taxon name}

\item{bp}{Import from \code{importBugphyzz}.}

\item{tax.id.type}{A character string. Either 'NCBI_ID' or 'Taxon_name'.}
}
\value{
A list of physiologies per taxa.
}
\description{
\code{getBugAnnotations} get all physiology annotations for one or more taxa.
}
\examples{

x <- getBugAnnotations(
    x = c('561', '562'), bp = importBugphyzz(), tax.id.type = 'NCBI_ID'
)
x

}
