% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugphyzz.R
\name{getBugphyzzSignatures}
\alias{getBugphyzzSignatures}
\title{Get Bugphyzz Signatures}
\usage{
getBugphyzzSignatures(
  df,
  tax.id.type = "NCBI_ID",
  tax.level = "mixed",
  evidence = c("asr", "inh", "exp", "tas", "nas", "igc"),
  frequency = c("unknown", "always", "usually", "sometimes"),
  min.size = 5
)
}
\arguments{
\item{df}{A data.frame}

\item{tax.id.type}{NCBI_ID or Taxon_name.}

\item{tax.level}{Domain to strain.}

\item{evidence}{asr, inh, exp, igc, nas, tas}

\item{frequency}{unknown, sometimes, usually, always}

\item{min.size}{Default 5.}
}
\value{
List of signatures
}
\description{
\code{getBugphyzzSignatures} convert data to signatures
}
\examples{

library(purrr)
bp <- importBugphyzz()
my_fun <- function(x) {
  getBugphyzzSignatures(
    df = x, tax.id.type = 'Taxon_name', tax.level = 'species', min.size = 5,
    frequency = 'always'
  )
}
o <- flatten(map(bp, my_fun))


}
