% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPhysiologies.R
\name{getPhysiologies}
\alias{getPhysiologies}
\title{Get physiologies}
\usage{
getPhysiologies(x, phys = physiologies(), taxid.type = "Taxon_name")
}
\arguments{
\item{x}{A valid NCBI ID or taxon name}

\item{phys}{A list of dataframes with physiologies imported with the \code{physiologies} function.
If not specified, all physiologies will be imported.}

\item{taxid.type}{The type of taxid. Valid options: Taxon_name or NCBI_ID.
Default is Taxon_name.}
}
\value{
A list of physiologies per taxa.
}
\description{
\code{getPhysiologies} get all physiology annotations for one or more taxa.
}
\examples{

phys <- physiologies()

taxa_names <- c('Escherichia coli', 'Staphylococcus aureus')
taxa_ids <- c('562', '1280')

phys_names <- getPhysiologies(taxa_names, phys)
phys_ids <- getPhysiologies(taxa_ids, phys, 'NCBI_ID')

}
