% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSignatures.R
\name{getSignatures}
\alias{getSignatures}
\title{Get signatures}
\usage{
getSignatures(
  df,
  tax.id.type = "Taxon_name",
  tax.level = "mixed",
  Frequency = c("unknown", "sometimes", "usually", "always"),
  Evidence = c("asr", "unknown", "igc", "asr"),
  min.size = 1,
  min = NULL,
  max = NULL
)
}
\arguments{
\item{df}{A data frame from bugphyzz.}

\item{tax.id.type}{Type of taxids. Check types with the \code{whichID} function.}

\item{tax.level}{A character vector with taxonomic ranks,
e.g., genus, species, strain.}

\item{Frequency}{A character vector with one or more values of: unknown,
sometimes, usually, always.}

\item{Evidence}{A character vector with one or more values of: unknown,
sometimes, usually, always.}

\item{min.size}{The minimun number of elements in a signatures.
Default is 1.}

\item{min}{Minimum threshold for attributes with continuous values. It's
ignored when creating signatures for categorical attributes.}

\item{max}{Maximum threshold for attributes with continuous values. It's
ignored when creating signatures for categorical attributes.}
}
\value{
A list of signatures. NULL if no signatures.
}
\description{
\code{getSignatures} creates a list of signatures from a bugphyzz
data frame.
}
\examples{

aer <- physiologies('aerophilicity')[[1]]
aer_sig <- getSignatures(aer)

ph <- physiologies('optimal ph')[[1]]
ph_sig <- getSignatures(ph)


}
