% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmtSignatures.R
\name{gmtSignatures}
\alias{gmtSignatures}
\title{Fetch signatures for easy GMT file creation}
\usage{
gmtSignatures(df)
}
\arguments{
\item{df}{A \code{data.frame} output by the \link{physiologies} function}
}
\value{
A list of attribute vectors
}
\description{
Fetch signatures for easy GMT file creation
}
\examples{
bugphyzz_GMT <- lapply(phys, get_signatures) \%>\%
unlist(recursive = F)
EnrichmentBrowser::writeGMT(bugphyzz_GMT, gmt.file = "data/bug_physiologies.gmt")
}
