% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugphyzz.R
\name{importBugphyzzNumeric}
\alias{importBugphyzzNumeric}
\title{Import bugphyzz data as a list of numeric data.frames}
\usage{
importBugphyzzNumeric(
  keyword = "all",
  version = "devel",
  force_download = FALSE
)
}
\arguments{
\item{keyword}{Name of the numeric attributes.
Use \code{showBugphyzzNumeric} to check valid values.
Default value is 'all'.}

\item{version}{Zenodo version of dataset. If 'devel' the data from
the bugphyzzExports repo on GitHub will be downloaded.}

\item{force_download}{use cache or a fresh download. Default is FALSE.}
}
\value{
A list of data.frames.
}
\description{
\code{importBugphyzzNumeric} imports numeric attributes as a list of
data.frames. Attributes are in separate data.frames since they can
have different units for filtering (e.g. Celsius degrees, ph, etc.)
}
\examples{

numeric <- importBugphyzzNumeric(version = 'devel', cache = TRUE)

}
