% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugphyzz.R
\name{importVal}
\alias{importVal}
\title{Import validation data}
\usage{
importVal()
}
\value{
A data.frame.
}
\description{
\code{importValidation} impots the result of the 10-fold cross-validation.
}
\examples{

val <- importValidation()

}
