% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_signatures.R
\name{make_signatures}
\alias{make_signatures}
\title{Makes taxid signatures for physiologies}
\usage{
make_signatures(df, taxids = "Taxon_name", rank = "all")
}
\arguments{
\item{df}{A \code{data.frame} output by the \link{physiologies} function}

\item{taxids}{identification of the bugs (Taxon name, NCBI ID, etc.)}

\item{rank}{Nothing yet}
}
\value{
a large list of character vectors
}
\description{
Makes taxid signatures for physiologies
}
\examples{
aero <- physiologies("aerophilicity")
x <- make_signatures(aero[[1]], taxids = "Taxon_name")
lapply(x, head)
x <- make_signatures(aero[[1]], taxids = "NCBI_ID")
lapply(x, head)
}
