% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical_summary_stats.R
\name{merge_ranks}
\alias{merge_ranks}
\title{Merge ranks vector with physiology dataset}
\usage{
merge_ranks(data, ranks)
}
\arguments{
\item{data}{a tibble object, converted from the output by the \link{physiologies} function}

\item{ranks}{a tibble object of NCBI ranks}
}
\value{
a tibble object
}
\description{
Merge ranks vector with physiology dataset
}
\examples{
physiologies <- physiologies()
\dontrun{
aer <- physiologies[["aerophilicity"]] \%>\%
as_tibble()

ncbi_vec <- lapply(physiologies, function(x) x\%>\% select(NCBI_ID)) \%>\%
unlist() \%>\%
as.numeric() \%>\%
unique()
ncbi_vec <- ncbi_vec[!is.na(ncbi_vec)]
ranks <- bugphyzz:::ncbiRank(ncbi_vec, ranks = c("phylum", "genus", "species"))

merge_ranks(aer, ranks)
}
}
\keyword{internal}
