% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical_summary_stats.R
\name{ncbi_stats}
\alias{ncbi_stats}
\title{Display number of taxa with/without NCBI ID in physiology dataset}
\usage{
ncbi_stats(data, phys = deparse(substitute(data)))
}
\arguments{
\item{data}{a tibble object, converted from the output by the \link{physiologies} function}

\item{phys}{a string vector of the physiology name}
}
\value{
a ggplot
}
\description{
Display number of taxa with/without NCBI ID in physiology dataset
}
\examples{
\dontrun{
physiologies <- physiologies()
aer <- physiologies[["aerophilicity"]] \%>\%
as_tibble()
ncbi_stats(aer, "aerophilicity")
}
}
\keyword{internal}
