% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy_functions.R
\name{ncbiRank}
\alias{ncbiRank}
\title{Get Taxonomy Ranks from NCBI Taxonomy IDs}
\usage{
ncbiRank(
  x,
  taxid = FALSE,
  ranks = c("superkingdom", "kingdom", "phylum", "class", "order", "family", "genus",
    "species")
)
}
\arguments{
\item{x}{A numeric vector or character vector of valid NCBI taxonomy IDs.}

\item{taxid}{A logical vector of length 1. If TRUE, taxonomy IDs (taxids) are returned instead of rank names.
Default is FALSE.}
}
\value{
A tibble. The first column, NCB_ID, contains the input taxonomy IDs.
}
\description{
\code{ncbiRank} retrieves taxonomy rank names (or IDs) at the kingdom, phylum, class,
order, family, genus, and species level for given NCBI taxonomy IDs.
}
\examples{
library(bugphyzz)
taxonomy_ids <- c(1578, 745, 562, 2)

# Example 1
rank_names <- ncbiRank(taxonomy_ids)
rank_names

# Example 2
rank_ids <- ncbiRank(taxonomy_ids, taxid = TRUE)
rank_ids

}
