% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical_summary_stats.R
\name{num_genera_per_attribute}
\alias{num_genera_per_attribute}
\title{Display number of genera per attribute in physiology dataset}
\usage{
num_genera_per_attribute(data, phys)
}
\arguments{
\item{data}{a tibble object, converted from the output by the \link{physiologies} function
and has been converted by the \link{fetch_ranks} function}

\item{phys}{a string vector of the physiology name}
}
\value{
a ggplot
}
\description{
Display number of genera per attribute in physiology dataset
}
\examples{
physiologies <- physiologies()
aer <- physiologies[["aerophilicity"]] \%>\%
as_tibble()
aer_ranks <- fetch_ranks(aer)
num_genera_per_attribute(aer_ranks, "aerophilicity")
}
\keyword{internal}
