% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physiologies2.R
\name{physiologies2}
\alias{physiologies2}
\title{Import physiologies}
\usage{
physiologies2(keyword = "all", remove_false = FALSE)
}
\arguments{
\item{keyword}{One or more values in a character vector. The values can
be checked with the \code{showPhys} function.}

\item{remove_false}{if TRUE, remove all attributes with FALSE values.
Default is FALSE (no filtering).}
}
\value{
A list of data frames.
}
\description{
\code{physiologies2} imports physiologies from Google spreadsheets.
}
\examples{

phys <- physiologies2('all')
aer <- physiologies2('aerophilicity')


}
