% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physiologies_list.R
\name{physiologies_list}
\alias{physiologies_list}
\title{Fetch list of available physiologies}
\usage{
physiologies_list()
}
\value{
a character vector of physiologies
}
\description{
Fetch list of available physiologies
}
\examples{
x <- physiologies_list()
}
