% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physiologies.R
\name{physiologiesList}
\alias{physiologiesList}
\title{List of available physiologies}
\usage{
physiologiesList()
}
\value{
A character vector with the names of the physiologies datasets.
}
\description{
\code{physiologiesList} prints the names of the available datasets provided by
the \code{\link{physiologies}} function.
}
\examples{
x <- physiologiesList()
}
