% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugphyzz.R
\name{print.signature}
\alias{print.signature}
\title{Method for printing signatures}
\usage{
\method{print}{signature}(x, ...)
}
\arguments{
\item{x}{A bugphyzz signature.}

\item{...}{Other arguments.}
}
\value{
The signature printed on screen.
}
\description{
Method for printing signatures
}
