% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical_summary_stats.R
\name{printGGlist}
\alias{printGGlist}
\title{Print function for ggplot list}
\usage{
printGGlist(gglist)
}
\arguments{
\item{gglist}{a list of ggplot}
}
\value{
printed ggplots
}
\description{
Print function for ggplot list
}
\examples{
\dontrun{
physiologies <- physiologies(keyword = "all") \%>\%
lapply(as_tibble)
ncbi_list <- mapply(function(.x, .y) bugphyzz:::ncbi_stats(.x, .y), physiologies, names(physiologies), SIMPLIFY = FALSE)
printGGlist(ncbi_list)
}
}
\keyword{internal}
