% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocFileCache.R
\name{taxonomyTable}
\alias{taxonomyTable}
\title{Taxonomy table}
\usage{
taxonomyTable(type = "name", update = FALSE)
}
\arguments{
\item{type}{A character vector of length 1. Options: "name" and "id".
If "name" the output will contain taxa names. If "id" the output will
contain NCBI IDs.}

\item{update}{If TRUE a new taxonomy table is created before importing.
If FALSE the taxonomy table currently stored in the user's cache is
imported.}
}
\value{
A tibble.
}
\description{
\code{taxonomyTable} imports a taxonomy table of all of the NCBI IDs
across BugPhyzz.
}
\details{
When the function is used for the first time, a taxonomy table is created
with \code{bugpyzz:::.createTaxonomyTable} before being imported. If NCBI IDs
have been added or removed from the BugPhyzz database, the taxonomy table
can be updated by setting the \code{update} parameter to TRUE. This will create
a new taxonomy table with \code{bugpyzz:::.createTaxonomyTable}.

Columns:
\itemize{
\item NCBI_ID: This column contains the original NCBI ID used to get the
taxonomy table. It can be used to join the taxonomy table with any bugphyzz
dataset.
\item rank: This column indicates the rank of the original NCBI_ID. Usually at
the genus, species, or strain level.
\item taxonomy ranks: There are other several columns such as superkingdom,
phylum, class, order, family, and genus filled with values for most NCBI IDs.
There are other columns such a species, strain, among a few others.
}
}
\examples{

\dontrun{
library(bugphyzz)
taxonomy_table <- taxonomyTable(type = "name", update = FALSE)
taxonomy_table
}

}
