% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugphyzz.R
\name{taxRanks}
\alias{taxRanks}
\title{Display taxonomic ranks}
\usage{
taxRanks()
}
\value{
A character vector
}
\description{
\code{taxRanks} display the names of the taxonomic ranks used in bugphyzz.
}
\examples{

taxRanks()

}
