% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_valid_headers.R
\name{valid_headers}
\alias{valid_headers}
\title{Valid headers of bugphyzz data sets}
\usage{
valid_headers()
}
\value{
A character vector with valid header names.
}
\description{
\code{valid_headers} contains the valid names and order of the mandatory
headers of a bugphyzz data set.
}
\examples{
x <- valid_headers()
x
}
