% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_valid_headers.R
\name{validHeaders}
\alias{validHeaders}
\title{Valid headers of bugphyzz data sets}
\usage{
validHeaders()
}
\value{
A character vector with valid header names.
}
\description{
\code{validHeaders} contains the valid names and order of the mandatory
headers of a bugphyzz data set.
}
\examples{
x <- validHeaders()
x
}
