% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugphyzz.R
\name{whichAttr}
\alias{whichAttr}
\title{Which Attributes}
\usage{
whichAttr(bp)
}
\arguments{
\item{bp}{A data.frame imported with \code{importBugphyzz}.}
}
\value{
A character vector.
}
\description{
\code{whichAttr} shows which attributes are present in a dataset imported
with \code{importBugphyzz}. Signatures created with
\code{getBugphyzzSignatures} would take these names.
}
\examples{

bp <- importBugphyzz()
whichAttr(bp)

}
