% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugphyzz.R
\name{whichAttrGrp}
\alias{whichAttrGrp}
\title{Which Attribute Groups}
\usage{
whichAttrGrp(bp)
}
\arguments{
\item{bp}{A data.frame imported with \code{importBugphyzz}.}
}
\value{
A character vector.
}
\description{
\code{whichAttrGrp} shows which attribute groups are present in a dataset
imported with \code{importBugphyzz}.
}
\examples{

bp <- importBugphyzz()
whichAttrGrp(bp)

}
