% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSignatures.R
\name{whichID}
\alias{whichID}
\title{Which IDs}
\usage{
whichID(df)
}
\arguments{
\item{df}{A data frame from bugphyzz.}
}
\value{
A character vector.
}
\description{
\code{whichID} gives valid values for the tax.id.type
argument of the \code\link{makeSignatures} function.
}
\examples{

aer <- physiologies('aerophilicity')[[1]]
whichID(aer)
}
