% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugphyzz.R
\name{bpScores.signature}
\alias{bpScores.signature}
\title{Function for getting bugphyzz scores from sigantures}
\usage{
\method{bpScores}{signature}(x, ...)
}
\arguments{
\item{x}{Object of class signature.}
}
\value{
A numeric vector with scores. Names are the Taxonomy IDs or Taxon
names.
}
\description{
Function for getting bugphyzz scores from sigantures
}
\examples{

bp <- importBugphyzz()
sig <- makeSignature(bp)
bpScores(sig[[1]])

}
