% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{checkColumnValues}
\alias{checkColumnValues}
\title{Check column values}
\usage{
checkColumnValues(
  column_name,
  dataset,
  dataset_name = NULL,
  quiet_success = TRUE
)
}
\arguments{
\item{column_name}{A character vector of length 1; name of the column.}

\item{dataset}{A data frame.}

\item{dataset_name}{A character vector of length 1; name of the dataset. Default is NULL.}

\item{quiet_success}{If FALSE, an error message is printed when no errors are found. Default is TRUE.}
}
\value{
\code{checkColumnValues} invisibly returns an error condition (subclass "invalid_column_values" or
"invalid_column_class"), and it also prints an error message if the column contains invalid values.
If \code{quiet_success} is FALSE, it prints a message indicating that no errors were found if that is
the case.
}
\description{
\code{checkColumnValues} checks if a single column in a data frame contains valid values as
according to the template.tsv file.
}
\examples{

\dontrun{

# Example 1

dataset_with_bad_values <- data.frame(
  NCBI_ID = letters[1:10],
  Genome_ID = 1:10,
  Accession_number = 1:10,
  Taxon_name = letters[1:10],
  Attribute = 1:10,
  Attribute_value = letters[1:10],
  Attribute_source = 1:10,
  Evidence = 1:10,
  Confidence_interval = 1:10,
  Note = 1:10
)

err1 = bugphyzz::checkColumnValues("NCBI_ID",  dataset_with_bad_values)
err1
err1$bad_values

err2 = bugphyzz::checkColumnValues("Attribute", dataset_with_bad_values)
err2
err2$bad_values

err3 = bugphyzz::checkColumnValues("Attribute_value", dataset_with_bad_values)
err3

# Example 2

aerophilicity <- bugphyzz::physiologies(keyword = "aerophilicity")[[1]]
aerophilicity_NCBI_ID_errors <- bugphyzz::checkColumnValues("NCBI_ID", aerophilicity) # no errors

}

}
\seealso{
\code{\link{checkColumnValues}};
\code{\link{checkColumnValuesDF}};
\code{\link{checkColumnValuesList}}

Other check functions: 
\code{\link{checkColumnValuesDF}()},
\code{\link{checkColumnValuesList}()},
\code{\link{checkRequiredColumnsList}()},
\code{\link{checkRequiredColumns}()}
}
\concept{check functions}
