% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{checkRequiredColumns}
\alias{checkRequiredColumns}
\title{Check required columns in a single bugphyzz dataset}
\usage{
checkRequiredColumns(dataset, dataset_name = NULL)
}
\arguments{
\item{dataset}{A data frame from bugphyzz}

\item{dataset_name}{Character of length 1; name of dataset. Default is NULL.}
}
\value{
\code{checkRequiredColumns} invisibly returns and error condition ("required_columns_missing" or
"required_columns_misplaced". It also returns an error message
if any of the required columns is missing or is not in the right order. If no errors are found,
\code{checkRequiredColumns} prints a message indicating that all required columns are present and
in the right order.
}
\description{
\code{checkRequiredColumns} checks if the required columns (see \code{\link{requiredColumns}}) are
present and in the right order in a single bugphyzz dataset. If a required column is missing or
is not in the right order, it prints a message indicating which columns must be added or reordered.
}
\examples{

\dontrun{

# Example 1 - dataset with missing required columns

dataset_with_missing_required_columns = data.frame(
  NCBI_ID = 1:10,
  Attribute = letters[1:10]
)
err1 <- bugphyzz:::checkRequiredColumns(dataset_with_missing_required_columns)
err1

# Example 2 - dataset with misplaced required columns

dataset_with_misplaced_required_columns = data.frame(
  NCBI_ID = 1:10,
  Taxon_name = letters[1:10],
  Genome_ID = 1:10,
  Accession_ID = 1:10,
  Attribute = 1:10,
  Attribute_value = 1:10,
  Attribute_source = 1:10,
  Evidence = 1:10,
  Note = 1:10,
  Confidence_interval = 1:10
)

err2 <- bugphyzz:::checkRequiredColumns(dataset_with_misplaced_required_columns)
err2

# Example 3 (dataset wit no errors)

dataset_with_required_columns_ok = data.frame(
  NCBI_ID = 1:10,
  Genome_ID = 1:10,
  Accession_ID = 1:10,
  Taxon_name = letters[1:10],
  Attribute = 1:10,
  Attribute_value = 1:10,
  Attribute_source = 1:10,
  Evidence = 1:10,
  Confidence_interval = 1:10,
  Note = 1:10
)

bugphyzz:::checkRequiredColumns(dataset_with_required_columns_ok)

# Example 4 - A bugphyzz dataset

oxygen <- bugphyzz:::physiologies(keyword = "aerophilicity")
bugphyzz:::checkRequiredColumns(x = oxygen[[1]], dataset_name = names(oxygen)[1])

}


}
\seealso{
\code{\link{requiredColumns}};
\code{\link{checkRequiredColumns}};
\code{\link{checkRequiredColumnsList}}

Other check functions: 
\code{\link{checkColumnValuesDF}()},
\code{\link{checkColumnValuesList}()},
\code{\link{checkColumnValues}()},
\code{\link{checkRequiredColumnsList}()}
}
\concept{check functions}
