% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute.R
\name{curationLinks}
\alias{curationLinks}
\title{Show links to curation spreadsheets}
\usage{
curationLinks(keyword = "all")
}
\arguments{
\item{keyword}{a character vector of attribute names desired. For the available
attributes, run bugphyzz::attribute_list(). Use "all" for all available attributes.}
}
\value{
a data.frame with attribute names and URLs
}
\description{
Show links to curation spreadsheets
}
\examples{
curationLinks()
curationLinks(keyword = "oxygen")
curationLinks(keyword = c("oxygen", "gram"))
}
