% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physiologies.R
\name{curationLinks}
\alias{curationLinks}
\title{Show links to curation spreadsheets}
\usage{
curationLinks(keyword = "all")
}
\arguments{
\item{keyword}{a character vector of physiologies desired. For the available
physiologies, run bugphyzz::physiologies_list(). Use "all" for all available physiologies.}
}
\value{
a data.frame with physiology names and URLs
}
\description{
Show links to curation spreadsheets
}
\examples{
curationLinks()
curationLinks(keyword = "aerophilicity")
curationLinks(keyword = c("aerophilicity", "gram stain"))
}
