% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fatty_acid_composition.R
\name{customLinks}
\alias{customLinks}
\title{Show links to custom link spreadsheet}
\usage{
customLinks(keyword = "all")
}
\arguments{
\item{keyword}{a character vector of custom links desired. For the available
custom links, run bugphyzz::custom_link(). Use "all" for all available custom links.}
}
\value{
a data.frame if fattyacidComposition from customlinks.tsv
}
\description{
Show links to custom link spreadsheet
}
\examples{
customLinks()
customLinks(keyword = "fatty_acid_composition")
}
