% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fattyAcidComposition.R
\name{customLinks}
\alias{customLinks}
\title{Custom links}
\usage{
customLinks(keyword = "all")
}
\arguments{
\item{keyword}{A character vector with the name of custom datasets.
Use "all" for all available custom links.}
}
\value{
A data frame with custom links.
}
\description{
\code{customLinks} returns a data frame with links to custom data sets not
not included in the \code{\link{physiologies}} function.
}
\examples{
customLinks(keyword = "all")
}
\keyword{internal}
