% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{.checkColumnValues}
\alias{.checkColumnValues}
\title{Check column values}
\usage{
.checkColumnValues(col, dat, dat_name = NULL, quiet_success = TRUE)
}
\arguments{
\item{col}{A character string indicating the name of the column.}

\item{dat}{A data frame imported through the bugphyzz functions.}

\item{dat_name}{A character string indicating the name of the dataset.
Default is NULL.}

\item{quiet_success}{If FALSE, an error message is printed when no errors
are found. Default is TRUE.}
}
\value{
invisibly returns an error condition
(subclass "invalid_column_values" or "invalid_column_class"),
and it also prints an error message if the column contains invalid values.
If \code{quiet_success} is FALSE an no errors were found, it prints a message
indicating that no errors were found.
}
\description{
\code{.checkColumnValues} checks if a single column in a data frame contains
valid values according to the extdata/template.tsv file.
}
\examples{

\dontrun{

x <- physiologies("aerophilicity")[[1]]
err <- tryCatch(bugphyzz:::.checkColumnValues(x), error = function(e) e)

}

}
\seealso{
\code{\link{.checkColumnValues}};
\code{\link{.checkColumnValuesDF}};
\code{\link{.checkColumnValuesList}}

Other check functions: 
\code{\link{.checkColumnValuesDF}()},
\code{\link{.checkColumnValuesList}()},
\code{\link{.checkRequiredColumnsDFList}()},
\code{\link{.checkRequiredColumnsDF}()}
}
\concept{check functions}
\keyword{internal}
