% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physiologies.R
\name{.importPhysiology}
\alias{.importPhysiology}
\title{Import physiology}
\usage{
.importPhysiology(x, remove_false = FALSE)
}
\arguments{
\item{x}{A row from the output of \code{curationLinks()}.}

\item{remvoe_false}{If TRUE, attributes with FALSE values are dropped.
Default is FALSE (all valeus included).}
}
\value{
A data frame.
}
\description{
\code{.importPhysiology} imports a physiology directly from one row of the
\code{curationLinks()} output.
}
\keyword{internal}
