% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physiologies.R
\name{.importPhysiology}
\alias{.importPhysiology}
\title{Import physiology}
\usage{
.importPhysiology(x, remove_false, full_source)
}
\arguments{
\item{x}{A row from the output of \code{curationLinks()}.}

\item{remove_false}{If TRUE, attributes with FALSE values are dropped.
Default is FALSE (all values included).}

\item{full_source}{if TRUE, the full source is displayed. Otherwise, an
abbreviated form.}
}
\value{
A data frame.
}
\description{
\code{.importPhysiology} imports a physiology directly from one row of the
\code{curationLinks} output.
}
\keyword{internal}
