% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{.stop_custom}
\alias{.stop_custom}
\title{Function for helping to generate custom stop (error) conditions}
\usage{
.stop_custom(subclass, message, call = NULL, ...)
}
\arguments{
\item{subclass}{A character string indicating the name of the custom error.}

\item{message}{A character string with an informative message.}

\item{call}{Include or not the function call that generates the error;
e.g., \code{sys.call(-1)}.}

\item{...}{Any other argument useful to identify the source of the error
and/or how to fix it.}
}
\value{
Error condition. Object of class "error", "condition",
and a custom error class name (subclass).
}
\description{
\code{.stop_custom} is a helper function for generating custom stop (error)
conditions. This function was taken from
Hadley Wickham's Advanced R book, 2nd edition
(\url{https://adv-r.hadley.nz/conditions.html#custom-conditions}).
This function should be used within custom stop functions.
}
\seealso{
\code{\link{.stop_custom}};
\code{\link{.stop_required_columns_missing}};
\code{\link{.stop_required_columns_misplaced}}
\code{\link{.stop_invalid_column_values}};
\code{\link{.stop_uncatalogued_column}};

Other custom stop functions: 
\code{\link{.stop_invalid_column_values}()},
\code{\link{.stop_required_columns_misplaced}()},
\code{\link{.stop_required_columns_missing}()},
\code{\link{.stop_required_missing_row_value}()},
\code{\link{.stop_uncatalogued_column}()}
}
\concept{custom stop functions}
\keyword{internal}
