% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{.stop_invalid_column_class}
\alias{.stop_invalid_column_class}
\title{Stop condition for invalid column class (in Attribute_value column)}
\usage{
.stop_invalid_column_class(col_class, dat_name = NULL, ...)
}
\arguments{
\item{col_class}{A character string containing the name of an invalid class.}

\item{dat_name}{A character string with the name of the dataset.}

\item{...}{Any other argument useful to identify the source of the error
and/or how to fix it.}
}
\value{
Error condition.
Object of class: "invalid_column_class", "error", "condition".
}
\description{
\code{.stop_invalid_column_class} generates an error condition if a column
(Attribute_value) is not logical or numeric.
}
\seealso{
\code{\link{.stop_custom}};
\code{\link{.stop_invalid_column_values}};
\code{\link{.stop_invalid_column_class}};
\code{\link{.stop_uncataloged_column}};
\code{\link{.checkColumnValues}}

Other custom stop functions: 
\code{\link{.stop_custom}()},
\code{\link{.stop_invalid_column_values}()},
\code{\link{.stop_required_columns_misplaced}()},
\code{\link{.stop_required_columns_missing}()},
\code{\link{.stop_uncataloged_column}()}
}
\concept{custom stop functions}
\keyword{internal}
