% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_functions.R
\name{.stop_invalid_column_values}
\alias{.stop_invalid_column_values}
\title{Stop condition for invalid values in a column of a bugphyzz dataset}
\usage{
.stop_invalid_column_values(
  col,
  n_rows,
  dataset_name = NULL,
  values = NULL,
  ...
)
}
\arguments{
\item{col}{Character of length 1; a single column name.}

\item{n_rows}{Integer of length 1; number of rows with invalid values.}

\item{dataset_name}{Character of length 1; dataset name.}

\item{values}{Optional. Invalid values in column. Only the first three will be printed.}

\item{...}{Any other argument useful to identify the source of the error and how to fix it.}
}
\value{
Error condition. Object of class "invalid_column_names", "error", "condition".
}
\description{
\code{.stop_invalid_column_values} generates an error condition if a column contains invalid values.
}
