% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{.stop_required_columns_missing}
\alias{.stop_required_columns_missing}
\title{Stop condition for missing required columns in a bugphyzz dataset}
\usage{
.stop_required_columns_missing(cols, dat_name = NULL, ...)
}
\arguments{
\item{cols}{A character vector containing the names of missing columns.}

\item{dat_name}{A character string indicating the name of the dataset.}

\item{...}{Any other argument useful to identify the source of the error
and/or how to fix it.}
}
\value{
Error condition.
Object of class "required_columns_missing", "error", "condition".
}
\description{
\code{.stop_required_columns_missing} generates an error condition if one or
more of the required columns is missing from a bugphyzz dataset.
The required columns can be printed with the \code{\link{.requiredColumns}}
function. This function should be used within the
\code{\link{.checkRequiredColumns}} function.
}
\seealso{
\code{\link{.stop_custom}};
\code{\link{.stop_required_columns_missing}};
\code{\link{.stop_required_columns_misplaced}}

Other custom stop functions: 
\code{\link{.stop_custom}()},
\code{\link{.stop_invalid_column_class}()},
\code{\link{.stop_invalid_column_values}()},
\code{\link{.stop_required_columns_misplaced}()},
\code{\link{.stop_uncataloged_column}()}
}
\concept{custom stop functions}
\keyword{internal}
