% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fattyAcidComposition.R
\name{fattyAcidComposition}
\alias{fattyAcidComposition}
\title{Function for shaping fatty acid composition from Google Sheets using tidyR}
\usage{
fattyAcidComposition()
}
\value{
a tidy data frame of fattyAcidComposition from customlinks.tsv
}
\description{
Function for shaping fatty acid composition from Google Sheets using tidyR
}
\examples{
library(bugphyzz)
fattyAcidComposition()
x <- fattyAcidComposition()
dplyr::glimpse(x)

}
