% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fattyAcidComposition.R
\name{fattyAcidComposition}
\alias{fattyAcidComposition}
\title{Import the fatty acid composition dataset}
\usage{
fattyAcidComposition()
}
\value{
A tidy data frame of fattyAcidComposition from spreadsheet_customlinks.tsv
}
\description{
\code{fattyAcidComposition} returns a data frame with data about the
fatty acid composition of microbes.
}
\examples{
fattyAcidComposition()
x <- fattyAcidComposition()
x

}
