% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_NCBI.R
\name{fill_NCBI}
\alias{fill_NCBI}
\title{Fill in NCBI IDs where the cells have NA}
\usage{
fill_NCBI(data)
}
\arguments{
\item{data}{a dataframe from the \link{physiologies} function}
}
\value{
a dataframe
}
\description{
Fill in NCBI IDs where the cells have NA
}
\examples{
\dontrun{
aer <- physiologies("aerophilicity")
aer_NCBI <- fill_NCBI(aer)
}
}
\keyword{internal}
