% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugphyzz.R
\name{getBugphyzzSignatures}
\alias{getBugphyzzSignatures}
\title{Get bugphyzz signatures}
\usage{
getBugphyzzSignatures(
  df,
  tax.id.type = "NCBI_ID",
  tax.level = "mixed",
  evidence = c("asr", "inh", "exp", "tas", "nas", "igc"),
  frequency = c("unknown", "rarely", "always", "usually", "sometimes"),
  min.size = 5
)
}
\arguments{
\item{df}{A data.frame}

\item{tax.id.type}{A character string. NCBI_ID or Taxon_name.}

\item{tax.level}{A character string indicating taxonomic level.
Valid options: domain, phylum, class,
order, family, genus, species, strain.}

\item{evidence}{A character string indicating type of evidence.
Valid option: asr, inh, exp, igc, nas, tas. See details for meaning of
each keyword.}

\item{frequency}{A cahracter string indicating frequency.
Valid options: unknown, sometimes, usually, always. See details for
meaning of each keyword.}

\item{min.size}{An integer. Minimum number of elements in a signature.}
}
\value{
Named list of signatures.
}
\description{
\code{getBugphyzzSignatures} convert a data.frame imported with bugphyzz
to signatures.
}
\examples{

## load purrr package (for managing lists)
library(purrr)

## Create helper function
sig_fun <- function(x) {
  getBugphyzzSignatures(
  df = x, tax.id.type = 'NCBI_ID', tax.level = 'species'
  )
}

## Create signatures of categorical or binary attributes
bp <- importBugphyzz()
sigs <- sig_fun(bp)
map(sigs, head)

## Create signatures of numeric attributes
bp_num <- importBugphyzzNumeric()
num_sigs <- flatten(map(bp_num, sig_fun))
lapply(num_sigs, head)

}
