% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSignatures.R
\name{getSignatures}
\alias{getSignatures}
\title{Get signatures}
\usage{
getSignatures(df, rank = "genus", id_type = "name", min = NULL, max = NULL)
}
\arguments{
\item{df}{A dataset from bugphyzz.}

\item{rank}{A character string with the taxonomy rank of the signatures.
Options: "superkingdom", "kingdom", "phylum", "class", "order", "family",
"species", "strain".}

\item{id_type}{A character string with the type of ID. Options: "Taxon_name", "NCBI_ID", "Accession_number", "Genome_ID".}

\item{min}{An integer. Minimum value. Only applies for attributes with
continuous values.}

\item{max}{An integer. Maximum value. Only applies for attributes with
continoues values.}
}
\value{
A list of microbe signatures
}
\description{
\code{getSignatures} creates a list of microbe signatures from a bugphyzz
dataset. The signatures are created from inherited ranks.
}
\examples{

library(bugphyzz)

# Make signatures of bacteria classified by oxygen utilization

aer <- physiologies("aerophilicity")[[1]]
aer_signatures <- getSignatures(df = aer, rank = "genus", id_type = "name")

# Make signature of bacteria growing in extreme temperatures

gt <- physiologies("growth temperature")[[1]]
gt_signatures <- getSignatures(df = gt, rank = "species", id_type = "name", min = 80)

}
