% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugphyzz.R
\name{getTaxonSignatures}
\alias{getTaxonSignatures}
\title{Get Taxon Signatures}
\usage{
getTaxonSignatures(tax, bp, ...)
}
\arguments{
\item{tax}{A valid NCBI ID or taxon name. If taxon name is used, the
tax_id_type = "Taxon_name" must also be used.}

\item{bp}{Import from \code{importBugphyzz}.}

\item{...}{Arguments passed to \code{makeSignatures}.}
}
\value{
A character vector with the names of the signatures for a taxon.
}
\description{
\code{getTaxonSignatures} get the names of all of the signatures for a taxon.
}
\examples{
taxid <- "562"
bp <- importBugphyzz()
sig_names_1 <- getTaxonSignatures("562", bp)
sig_names_2 <- getTaxonSignatures("Escherichia coli", bp, tax_id_type = "Taxon_name")

}
