% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugphyzz.R
\name{importBugphyzz}
\alias{importBugphyzz}
\title{Import bugphyzz}
\usage{
importBugphyzz(
  version = "devel",
  force_download = FALSE,
  v = 0.5,
  exclude_rarely = TRUE
)
}
\arguments{
\item{version}{Character string. Default is 'devel'
(current file on the GitHub repo waldronlab/bugphyzzExports).}

\item{force_download}{Logical value. Force a fresh download of the data or
use the one stored in the cache (if available). Default is FALSE.}

\item{v}{Validation value. Default 0.5.}

\item{exclude_rarely}{Default is TRUE. Exclude values with
Frequency == FALSE.}
}
\value{
A list of data frames.
}
\description{
\code{importBugphyzz} imports bugphyzz annotations as a list of
data.frames (1 per physiology/Attribute)
}
\examples{

bp <- importBugphyzz()
names(bp)

}
