% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugphyzz.R
\name{importBugphyzz}
\alias{importBugphyzz}
\title{Import bugphyzz data as a single data.frame}
\usage{
importBugphyzz(keyword = "all", version = "devel", force_download = FALSE)
}
\arguments{
\item{keyword}{The name of the physiology/attribute. Default is 'all'.}

\item{version}{Version to download. Default is 'devel'.}

\item{force_download}{Use cache or use a fresh download. Default is FALSE.}
}
\value{
A list of data.frames (tibbles).
}
\description{
\code{importBugphyzz} imports bugphyzz.
}
\examples{

bp <- importBugphyzz()

}
