% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugphyzz.R
\name{importBugphyzz}
\alias{importBugphyzz}
\title{Import bugphyzz}
\usage{
importBugphyzz(version = "devel", force_download = FALSE, v = 0.5)
}
\arguments{
\item{version}{Character string. The version to download. Default is 'devel'
(current file on the GitHub repo waldronlab/bugphyzzExports).}

\item{force_download}{Logical value. Force a fresh download of the data or
use the one stored in the cache (if available). Default is FALSE.}

\item{v}{Validation value. Deafult 0.5.}
}
\value{
A data.frame.
}
\description{
\code{importBugphyzz} imports bugphyzz as a single data.frame. It includes
attributes with categorical and binary (boolean) attributes.
}
\examples{

bp <- importBugphyzz()
names(bp)

}
