% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugphyzz.R
\name{importBugphyzz}
\alias{importBugphyzz}
\title{Import bugphyzz as a single data.frame (or several?}
\usage{
importBugphyzz(keyword = "all", version = "devel", cache = TRUE)
}
\arguments{
\item{keyword}{The name of the physiology/attribute. Default is 'all'.}

\item{version}{Version to download. Default is 'devel'.}

\item{cache}{Use cache or use a fresh download.}
}
\value{
A list of data.frames (tibbles).
}
\description{
\code{importBugphyzz} imports bugphyzz.
}
\examples{

bp <- importBugphyzz()

}
