% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugphyzz.R
\name{importBugphyzzNumeric}
\alias{importBugphyzzNumeric}
\title{Import bugphyzz (numeric/continuous data)}
\usage{
importBugphyzzNumeric(
  keyword = "all",
  version = "devel",
  force_download = FALSE
)
}
\arguments{
\item{version}{Character string. The version to download. Default is 'devel'
(current file on the GitHub repo waldronlab/bugphyzzExports).}

\item{force_download}{Logical value. Force a fresh download of the data or
use the one stored in the cache (if available). Default is FALSE.}
}
\value{
A list of data.frames.
}
\description{
\code{importBugphyzzNumeric} imports a list of data.frames containing
numeric attributes. In this case, the attributes are in separate
data.frames because they can have different units or scales
(e.g. Celsius degrees, ph, etc.)
}
\examples{

bp_num <- importBugphyzzNumeric()

## Check available numeric attributes
names(bp_num)

## Select growth temperature
gt <- bp_num[['growth temperature']]

## Get taxa that grows better betwen 0 and 25 Celsius degrees
sub_gt <- gt[which(gt$Attribute_value_min >= 0 & gt$Attribute_value_max <= 25),]

## Creat signature at the genus level
sigs <- getBugphyzzSignatures(sub_gt, tax.id.type = 'Taxon_name', tax.level = 'genus')
head(sigs[[1]])

}
