% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_signatures.R
\name{make_signatures}
\alias{make_signatures}
\title{Makes taxid signatures for physiologies}
\usage{
make_signatures(physiologies = "all", taxids = "Taxon_name", rank = "all")
}
\arguments{
\item{physiologies}{the desired physiology}

\item{taxids}{identification of the bugs (Taxon name, NCBI ID, etc.)}

\item{rank}{Nothing yet}
}
\value{
a large list of character vectors
}
\description{
Makes taxid signatures for physiologies
}
\examples{
x <- make_signatures("aerophilicity")
head(x[["aerobic"]])
}
