% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSignatures.R
\name{makeSignatures}
\alias{makeSignatures}
\title{Makes taxid signatures for physiologies}
\usage{
makeSignatures(
  df,
  taxids = "Taxon_name",
  taxa_rank = "genus",
  min = NULL,
  max = NULL
)
}
\arguments{
\item{df}{A \code{data.frame} output by the \link{physiologies} function}

\item{taxids}{identification of the bugs (Taxon name, NCBI ID, etc.)}

\item{taxa_rank}{Nothing yet}

\item{min}{An integer. Minimum value. Only applies for attributes with
continuous values.}

\item{max}{An integer. Maximum value. Only applies for attributes with
continoues values.}
}
\value{
a large list of character vectors
}
\description{
\code{makeSignatures}  reates a list of microbe signatures from a bugphyzz
dataset.
}
\examples{
aero <- physiologies("aerophilicity")
x <- makeSignatures(aero[[1]], taxids = "Taxon_name")
lapply(x, head)
x <- makeSignatures(aero[[1]], taxids = "NCBI_ID")
lapply(x, head)
}
