% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSignatures.R
\name{makeSignatures}
\alias{makeSignatures}
\title{Make signatures}
\usage{
makeSignatures(keyword = NULL, ...)
}
\arguments{
\item{keyword}{A valid keyword as determined by \code{physiologiesList()}.}

\item{...}{Parameters used for the \code{getSignature} function.}
}
\value{
A list of signatures. NULL if no signatures.
}
\description{
\code{makeSignatures} creates a list of signatures.
}
\examples{
aer_sig <- makeSignatures('aerophilicity')
ph_sig <- makeSignatures('optimal ph')

}
