% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSignatures.R
\name{makeSignatures}
\alias{makeSignatures}
\title{Make signatures}
\usage{
makeSignatures(
  df,
  tax_id_type = "NCBI_ID",
  tax_level = "mixed",
  evidence = c("EXP", "ASR", "Unknown"),
  ci = 0.8,
  include_unknown_ci = FALSE,
  min_size = 5
)
}
\arguments{
\item{df}{A data.frame imported with bugphyzz.}

\item{tax_id_type}{Type of ID. Options: NCBI_ID, Taxon_name.}

\item{tax_level}{Taxonomy level. Options: superkingdom, kingom, phylum,
class, order, family, genus, species, strain, mixed (all taxonomies).
Default is mixed.}

\item{evidence}{Type of evidence of the assigned attribute values to a
given taxon. Options: EXP, experimental; ASR, ancestral state reconstruction.
Default is both EXP and ASR.}

\item{ci}{A double indicating how likely is that a given
organism has that attribute value. Threshold options: 1, always; 0.8,
Usually; 0.5, sometimes; 0.3, Occasionally.
Values below 0.3 are not allowed. Default is 0.8.}

\item{include_unknown_ci}{Include taxa whose attribute value confidence
interval is unknown.}

\item{min_size}{The minimal size of the returned microbial signatures.
Default is 1.}
}
\value{
A list of microbial signatures
}
\description{
\code{makeSignatures} converts a bugphyzz dataset into a list of microbial
signatures.
}
\examples{

aer <- physiologies("aerophilicity")[[1]]
aer_signatures <- makeSignatures(aer)
lapply(aer, head)

}
