% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bugphyzz.R
\name{makeSignatures}
\alias{makeSignatures}
\title{Make bugs signatures}
\usage{
makeSignatures(
  dat,
  tax_id_type = "NCBI_ID",
  tax_level = "mixed",
  evidence = c("exp", "igc", "tas", "nas", "tax", "asr"),
  frequency = c("always", "usually", "sometimes", "unknown"),
  min_size = 10,
  min = NULL,
  max = NULL
)
}
\arguments{
\item{dat}{A data.frame.}

\item{tax_id_type}{A character string. Valid options: NCBI_ID, Taxon_name.}

\item{tax_level}{A character vector. Taxonomic rank. Valid options:
kingdom, phylum, class, order, family, genus, species, strain.
They can be combined. "mixed" is equivalent to select all valid ranks.}

\item{evidence}{A character vector. Valid options: exp, igc, nas, tas, tax, asr.
They can be combined. Default is all.}

\item{frequency}{A character vector. Valid options: always, usually,
sometimes, rarely, unknown. They can be combiend. Default value is all but
rarely.}

\item{min_size}{Minimun number of bugs in a signature. Default is 10.}

\item{min}{Minimum value inclusive. Only for numeric attributes. Default is NULL.}

\item{max}{Maximum value inclusive. Only for numeric attributes. Default is NULL.}
}
\value{
A list of character vector with the IDs of the bugs.
}
\description{
\code{makeSignatures} Creates signatures for a list of bugphyzz
data.frames imported with \code{importBugphyzz}
}
\examples{

bp <- importBugphyzz()
sigs <- lapply(bp, makeSignatures)
sigs <- purrr::list_flatten(sigs)

}
