% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSignatures.R
\name{makeSignatures}
\alias{makeSignatures}
\title{Make signatures}
\usage{
makeSignatures(
  keyword,
  tax.id.type,
  tax.level = "mixed",
  Frequency = c("unknown", "sometimes", "usually", "always"),
  Evidence = c("unknown", "exp", "igc", "inh", "exp"),
  min.size = 1,
  min = NULL,
  max = NULL,
  num = NULL,
  remove_false = FALSE
)
}
\arguments{
\item{keyword}{One or more values in a character vector. The values can
be checked with the \code{showPhys} function.}

\item{tax.id.type}{A character string. NCBI_ID or Taxon_name.}

\item{tax.level}{A taxonomy level. Check valid ranks with \code{validRanks()}.
Default is 'mixed', which uses all valid ranks.}

\item{Frequency}{A character string. Possible values are 'always',
'sometimes', 'usually', 'unknown'.}

\item{Evidence}{A character string. Possible values are 'exp', 'igc', 'inh',
'asr', 'unknown'.}

\item{min.size}{Minimum size of the signatures. Default is 1.}

\item{min}{Minimum for range and numeric attributes.}

\item{max}{Maximum for range and numeric attributes.}

\item{num}{Exact number for numeric attributes.}

\item{remove_false}{Whether include the creation of signatures with FALSE
values or not. Default is TRUE, i.e., FALSE values are not included.}
}
\value{
List of signatures
}
\description{
\code{makeSignatures} is a wrapper of \code{physiologies} followed by
\code{getSignatures}.
}
\examples{

bugphyzz_sigs <- makeSignatures('all', 'Taxon_name')


}
