% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSignatures.R
\name{makeSignatures}
\alias{makeSignatures}
\title{Make signatures}
\usage{
makeSignatures(
  keyword = NULL,
  tax.id.type = "Taxon_name",
  tax.level = "mixed",
  Frequency = c("unknown", "sometimes", "usually", "always"),
  Evidence = c("asr", "unknown", "igc", "inh"),
  min.size = 1,
  min = NULL,
  max = NULL
)
}
\arguments{
\item{keyword}{A valid keyword as determined by \code{physiologiesList()}.}

\item{tax.id.type}{Type of taxids. Check types with the \code{.whichID} function.}

\item{tax.level}{A character vector with taxonomic ranks,
e.g., genus, species, strain.}

\item{Frequency}{A character vector with one or more values of: unknown,
sometimes, usually, always.}

\item{Evidence}{A character vector with one or more values of: unknown,
sometimes, usually, always.}

\item{min.size}{The minimun number of elements in a signatures.
Default is 1.}

\item{min}{Minimum threshold for attributes with continuous values. It's
ignored when creating signatures for categorical attributes.}

\item{max}{Maximum threshold for attributes with continuous values. It's
ignored when creating signatures for categorical attributes.}
}
\value{
A list of signatures. NULL if no signatures.
}
\description{
\code{makeSignatures} creates a list of signatures.
}
\examples{
aer_sig <- makeSignatures('aerophilicity')
ph_sig <- makeSignatures('optimal ph')

}
\seealso{
\code{\link{getSignatures}}
}
