% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physiologies.R
\name{physiologies}
\alias{physiologies}
\title{Import physiologies}
\usage{
physiologies(keyword = "all", full_source = FALSE)
}
\arguments{
\item{full_source}{If TRUE, full source is provided. If FALSE, a
shortened name of the source is provided. Default if FALSE.}

\item{Character}{vector with one or more keywords. Valid keyboard
values can be checked with \code{showPhys}.}
}
\value{
A list of data.frames.
}
\description{
\code{physiologies} imports data about physiologies and other traits
stored in Google spreadsheets and the main bacdive file.
}
\examples{

phys <- physiologies('all')
aer <- physiologies('aerophilicity')

}
