% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physiologies.R
\name{physiologies}
\alias{physiologies}
\title{Physiologies}
\usage{
physiologies(keyword = "all", full_source = FALSE)
}
\arguments{
\item{keyword}{Character vector with one or more valid keywords.
Valid keyboards can be checked with \code{showPhys}. If 'all', all
physiologies are imported.}

\item{full_source}{Logical. If \code{TRUE}, the Attribute_source column will
contain full source information. If \code{FALSE}, the Attribute_source column will
contain shortened versions of the sources. Default is \code{FALSE}.}
}
\value{
A list of data.frames in tidy format.
}
\description{
\code{physiologies} imports data from the
Google spreadsheets (https://drive.google.com/drive/folders/1i2UAolVWAYa7UnETNnCs0BDWjKPp3ev5).
}
\examples{
l <- physiologies('all')
df <- physiologies('aerophilicity')[[1]]

}
