% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physiologies.R
\name{physiologies}
\alias{physiologies}
\title{Import physiologies}
\usage{
physiologies(keyword = "all", remove_false = FALSE, full_source = TRUE)
}
\arguments{
\item{keyword}{One or more values in a character vector. The values can
be checked with the \code{showPhys} function.}

\item{remove_false}{if TRUE, remove all attributes with FALSE values.
Default is FALSE (no filtering).}

\item{full_source}{If TRUE, full source is provided. If FALSE, a
shortened name of the source.}
}
\value{
A list of data frames.
}
\description{
\code{physiologies} imports physiologies from Google spreadsheets.
}
\examples{

phys <- physiologies('all')
aer <- physiologies('aerophilicity')


}
