% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physiologies.R
\name{physiologies}
\alias{physiologies}
\title{Fetch physiological data sheets}
\usage{
physiologies(keyword = "all")
}
\arguments{
\item{keyword}{a character vector of physiologies (see \code{\link{curationLinks}}). For the available
physiologies, run bugphyzz::physiologies_list()}
}
\value{
a large list of tidy data.frames
}
\description{
Fetch physiological data sheets
}
\examples{
x <- physiologies()
head(x[[1]])

y <- physiologies(c("gram", "aerophilicity"))
head(y[[1]])
}
