% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physiologies_list.R
\name{physiologies_list}
\alias{physiologies_list}
\title{Fetch list of available physiological physiologies names}
\usage{
physiologies_list()
}
\value{
a character vector of physiological physiologies names
}
\description{
Fetch list of available physiological physiologies names
}
\examples{
x <- physiologies_list()
}
