% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physiologies.R
\name{showPhys}
\alias{showPhys}
\title{List of available physiologies}
\usage{
showPhys()
}
\value{
A character vector with the names of the physiologies datasets.
}
\description{
\code{showPhys} prints the names of the available datasets provided by
the \code{\link{physiologies}} function.
}
\examples{
x <- showPhys()
}
