% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physiologies.R
\name{showPhys}
\alias{showPhys}
\title{Show list of available physiologies}
\usage{
showPhys(which_names = "all")
}
\arguments{
\item{which_names}{A character string. Options: 'all' (default),
'spreadsheets', 'bacdive'.}
}
\value{
A character vector with the names of the physiologies.
}
\description{
\code{showPhys} prints the names of the available physiologies that can be
imported with the \code{\link{physiologies}} function.
}
\examples{
showPhys()
showPhys('bacdive')
showPhys('spreadsheets')

}
